third_party/blink/web_tests/clipboard/async-clipboard/async-svg-datatransfer-read-write.tentative.https.html

<!doctype html>
<meta charset="utf-8">
<title>DataTransfer items read unsupported image/svg+xml types test.</title>
<body>Body needed for test_driver.click()</body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="resources/user-activation.js"></script>
<script>
'use strict';

promise_test(async t => {
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');
  const textInput = '<svg> <circle cx="50" cy="50" r="40"/> </svg>';
  const blobInput = new Blob([textInput], {type: 'image/svg+xml'});
  const clipboardItem = new ClipboardItem({'image/svg+xml': blobInput});
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItem]);
  // Write supported and unsupported types to clipboard.
  document.addEventListener('paste', event => {
    assert_equals(event.clipboardData.items.length, 0);
    event.preventDefault();
  });
  document.execCommand('paste');
}, 'Verify that dataTransfer items removes unsupported SVG type.');
</script>
